GH-382: [refactor] Update Kafka events dependency to version 1.0.6 and adjust imports in OrchestrationServiceConsumer and OrchestrationServiceConsumerImpl
