+++ Microservices/email-service/src/main/java/app/sportahub/emailservice/service/kafka/consumer/OrchestrationServiceConsumer.java
@@ -1,6 +1,7 @@
 package app.sportahub.emailservice.service.kafka.consumer;
 
-import app.sportahub.kafka.events.ForgotPasswordSendEmailEvent;
+
+import app.sportahub.kafka.events.forgotpassword.ForgotPasswordSendEmailEvent;
 
 public interface OrchestrationServiceConsumer {
