+++ Microservices/email-service/src/main/java/app/sportahub/emailservice/service/kafka/consumer/OrchestrationServiceConsumerImpl.java
@@ -1,7 +1,7 @@
 package app.sportahub.emailservice.service.kafka.consumer;
 
-import app.sportahub.kafka.events.ForgotPasswordEvent;
-import app.sportahub.kafka.events.ForgotPasswordSendEmailEvent;
+import app.sportahub.kafka.events.SportaKafkaEvents;
+import app.sportahub.kafka.events.forgotpassword.ForgotPasswordSendEmailEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.kafka.annotation.KafkaListener;
@@ -13,7 +13,7 @@ import org.springframework.stereotype.Service;
 public class OrchestrationServiceConsumerImpl implements OrchestrationServiceConsumer {
 
     @Override
-    @KafkaListener(topics = ForgotPasswordEvent.SEND_EMAIL_TOPIC, groupId ="OrchServiceKafkaConsumer")
+    @KafkaListener(topics = SportaKafkaEvents.SEND_EMAIL_TOPIC, groupId ="OrchServiceKafkaConsumer")
     public void listenForForgotPasswordSendEmailEvent(ForgotPasswordSendEmailEvent forgotPasswordSendEmailEvent){
         log.info("OrchestrationServiceConsumerImpl::ForgotPasswordSendEmailEvent: received forgot password request with email: {}", forgotPasswordSendEmailEvent.getEmail());
     }
